Abstract classname logic to a helper file to keep pb_forms_helper uncluttered
