+++ playbook/app/pb_kits/playbook/pb_form/form.rb
@@ -26,9 +26,6 @@ module Playbook
           loading: loading,
           validate: validate,
         }.merge(prop(:options) || prop(:form_system_options) || {})
-
-        props = options[:props] || {}
-        options.merge(props)
       end
     end
   end
